Feature article and comment

+++ server-nestjs/src/about-article/article/article.service.ts
@@ -1,11 +1,10 @@
-import { ArticleRepository } from './article.repository';
-import { CreateArticleDto } from './dtos/article.create.dto';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { AvatarService } from 'src/about-user/avatar/avatar.service';
-import { ArticleEntity } from './article.entity';
 import { DeleteResult, UpdateResult } from 'typeorm';
+import { ArticleEntity } from './article.entity';
+import { ArticleRepository } from './article.repository';
+import { CreateArticleDto } from './dtos/article.create.dto';
 import { UpdateArticleDto } from './dtos/article.update.dto';
-import { AvatarEntity } from 'src/about-user/avatar/avatar.entity';
 
 @Injectable()
 export class ArticleService {
@@ -22,30 +21,32 @@ export class ArticleService {
     return await this.articleRepository.createArticle(createArticleDto, avatar);
   }
 
-  async findArticleByArticleId(
-    articleId: number,
-    avatarId: number,
-  ): Promise<ArticleEntity> {
-    const article =
-      await this.articleRepository.findArticleByArticleId(articleId);
+  async findArticleById(articleId: number): Promise<ArticleEntity> {
+    const article = await this.articleRepository.findArticleById(articleId);
     if (!article) {
       throw new NotFoundException('해당하는 게시글을 찾을 수 없습니다.');
     }
-    // 게시글이 공개인지 비공개인지 확인 후 게시글의 주인과 로그인한 아바타와 동일한지 비교
-    return await this.articleRepository.findArticleByArticleId(articleId);
+    return article;
   }
 
   async findAllComments(): Promise<ArticleEntity[]> {
     return await this.articleRepository.findAllComments();
   }
 
-  async updateArticleByArticleId(
+  async updateArticle(
     articleId: number,
     updateArticleDto: UpdateArticleDto,
     avatarId: number,
   ): Promise<UpdateResult> {
-    const article = await this.findArticleByArticleId(articleId, avatarId);
-    return await this.articleRepository.updateArticleByArticleId(
+    const article = await this.findArticleById(articleId);
+    const ownerId = article.avatar.avatarId;
+
+    const isOwner = await this.isOwner(ownerId, avatarId);
+    if (!isOwner) {
+      throw new NotFoundException('게시글 작성자만 수정할 수 있습니다.');
+    }
+
+    return await this.articleRepository.updateArticle(
       articleId,
       updateArticleDto,
     );
@@ -55,7 +56,18 @@ export class ArticleService {
     articleId: number,
     avatarId: number,
   ): Promise<DeleteResult> {
-    const article = this.findArticleByArticleId(articleId, avatarId);
+    const article = await this.findArticleById(articleId);
+    const ownerId = article.avatar.avatarId;
+
+    const isOwner = await this.isOwner(ownerId, avatarId);
+    if (!isOwner) {
+      throw new NotFoundException('게시글 작성자만 수정할 수 있습니다.');
+    }
+
     return await this.articleRepository.deleteArticle(articleId);
+  }
+
+  async isOwner(ownerId: number, avatarId: number) {
+    return ownerId === avatarId;
   }
 }
